parse by float to preserve decimal places (#2383)
* Add support for collapsible callouts

* aeolian comments

* fix sizing

* Fix bug with collapsed form values

* use ipfs in creation and load ipfs images in project logo

* fix bad rebase

* Add loading state for go to project when clicked

* Fix Specific Funding Target showing infinite when switching from inifinite to specific

* Change amount form item based on amountType

* Hide token minting callout on disabled

* Fix some spacing on review deploy

* parse by float to preserve decimal places

+++ src/components/Create/components/Allocation/AddEditAllocationModal.tsx
@@ -345,7 +345,7 @@ const DistributionAmountInput = ({
   if (mode === 'percentage') {
     return (
       <NumberSlider
-        sliderValue={value?.percentage ? parseInt(value.percentage) : 0}
+        sliderValue={value?.percentage ? parseFloat(value.percentage) : 0}
         onChange={percentage =>
           onAmountInputChange({ percentage: percentage?.toString() })
         }
